+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: aletude- <aletude-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/17 17:14:29 by aletude-          #+#    #+#             */
-/*   Updated: 2025/07/25 16:24:34 by aletude-         ###   ########.fr       */
+/*   Updated: 2025/07/31 12:10:49 by aletude-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,10 +35,10 @@ int				ft_isascii(int c);
 int				ft_isalnum(int c);
 int				ft_isdigit(int c);
 int				ft_isprint(int c);
-int				ft_strlen(const char *str);
+size_t			ft_strlen(const char *str);
 void			*ft_memset(void *s, int c, size_t n);
 void			*ft_memcpy(void *dest, const void *src, size_t n);
-unsigned int	ft_strlcat(char *dest, const char *src, unsigned int size);
+size_t			ft_strlcat(char *dest, const char *src, size_t size);
 size_t			ft_strlcpy(char *dest, const char *src, size_t size);
 void			*ft_memmove(void *dest, const void *src, size_t n);
 int				ft_toupper(int c);
